backend/requirements.txt
Django==5.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
channels==4.2.2
channels_redis==4.2.1
daphne==4.0.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
uvicorn==0.34.2
dj-database-url==2.1.0
